max_new_tokens as a parameter to answer_question and default 256
max_new_tokens is added as a parameter to answer_question() and default is set to 256 instead of 512. Reasons: 1) more control, 2) at least in test generations I found about 200 wasn't interrupted (100 was small) ; 3) in order to reduce the delay when the model enters in loops generating .............. and taking a lot of time. Also: can some sort of an interrupt be added, either with a callback like in GPT4All, also a constraint when capturing a repetition of tokens etc., repetition_penalty etc.? Can we add a streaming mode? Thanks.