added moondream_batch initialization
Allocates memory for elements of moondream_batch struct.
The alloc_embd boolean arg controls whether memory is allocated for
tokens or for the token embeddings. When alloc_embd is true, memory is
allocated for the token embeddings (float vectors), when it is false,
memory is allocated for the tokens (integers).