init backend scheduler and use it to alloc graph tensors
ggml_backend_sched_alloc_graph uses the backend scheduler
(ggml_backend_sched_t) to set the backend buffers for all of the tensors
in the cgraph. The scheduler is initialized with the worst-case graph
(n_tokens = n_ctx) during the moondream_context initialization. With
this setup we can now populate the inp_tokens tensor with the token ids
from the batch struct without getting an error saying that the
inp_tokens buffer is not set.

Regarding the deleted comment about the build callback being where the
backend is set... this is only true for specific tensors, the rest are
set by ggml_backend_sched_alloc_graph.