generate tokens in a loop and log them to the console
The generation loop can now run with the graph compute call and will not
crash (at least for 30 or so steps; I haven't tested further). This
means we can now fetch the logits and sample them to generate a new
token, unfortunately the output is junk. For example, with the prompt
"Describe this image.", the response is
"ribeĠasĠasribingrr:Ġasribe.riberibe.ribeRriberiberibe" after 17 steps.
This might be due to the lack of image tokens (there is no mmproj right
now), or something wrong with the current implementation like incorrect
tokenization. Testing with base phi2 weights would rule out the former.

Summary of changes:

- Added some additional error checking where ggml API calls can fail.
- Hardcoded rope and yarn parameters (these could be wrong and might be
  one reason for the garbage output).
- Switch type_k and type_v to GGML_TYPE_F32. Phi2 uses F32 for attention
  either way so I'm not sure if these paramters make any difference.
- Set mctx.n_outputs to 1 after initialization. This should give us 1
  set of logits per step, or in other words, 1 new token per step.
- Set batch token ids to the ids of the tokenized prompt.
- Implemented hacky logit extraction and sampling. The sampled_token_id
  is equal to the argmax of the logits. Proper sampling will have to be
  implemented later. Logits should also be extracted to a persistent
  buffer instead of one that is allocated and de-allocated each step.
- mctx.n_sample and batch.n_tokens are now incremented after each step.
- Removed ggml_backend_cpu_set_n_threads call from generation loop
  because it is already set when moondream_context is initialized.