separate mmproj and lm into different source files
It was getting a bit confusing to navigate a single file with the code
for two different models. Now the language model code is in
lm.cpp/lm.hpp and the mmproj code is in mmproj.cpp/mmproj.hpp so the lm
won't get in the way when working on the mmproj and vice versa.
Additionaly, all helper functions and shared macros are now in
helper.cpp and helper.hpp.

moondream.cpp now just has the program entry point and the api
state/functions.